requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
annotated-types==0.7.0
    # via pydantic
pydantic==2.11.3
    # via -r requirements.in
pydantic-core==2.33.1
    # via pydantic
sseclient-py==1.8.0
    # via -r requirements.in
typing-extensions==4.13.2
    # via
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.0
    # via pydantic
